+++ samples/emoji-search/schema/src/main/kotlin/com/example/redwood/emojisearch/schema.kt
@@ -17,7 +17,6 @@ package com.example.redwood.emojisearch
 
 import app.cash.redwood.layout.RedwoodLayout
 import app.cash.redwood.lazylayout.RedwoodLazyLayout
-import app.cash.redwood.schema.Default
 import app.cash.redwood.schema.Modifier
 import app.cash.redwood.schema.Property
 import app.cash.redwood.schema.Schema
@@ -42,14 +41,11 @@ interface EmojiSearch
 @Widget(1)
 data class TextInput(
   @Property(1)
-  @Default("TextFieldState()")
-  val state: TextFieldState,
+  val state: TextFieldState = TextFieldState(),
   @Property(2)
-  @Default("\"\"")
-  val hint: String,
+  val hint: String = "",
   @Property(3)
-  @Default("null")
-  val onChange: ((TextFieldState) -> Unit)?,
+  val onChange: ((TextFieldState) -> Unit)? = null,
 )
 
 @Widget(2)
@@ -60,7 +56,7 @@ data class Text(
 @Widget(3)
 data class Image(
   @Property(1) val url: String,
-  @Property(2) @Default("null") val onClick: (() -> Unit)? = null,
+  @Property(2) val onClick: (() -> Unit)? = null,
 )
 
 @Modifier(-4_543_827) // -4_543_827 is a reserved tag.